requirements.txt
fastapi
uvicorn[standard]
httpx
python-multipart
django
gunicorn
uvicorn
python-dotenv
